refactor(server): fix obvious naming issues
