+++ packages/server/src/users/favourites/api/UserFavouriteRestaurantsController.ts
@@ -2,13 +2,13 @@ import { Res } from '@nestjs/common';
 import { plainToInstance } from 'class-transformer';
 import { Response } from 'express';
 
-import { ApiController, ApiList, createPaginationLink, Pagination, PaginationQuery, Url } from '../../shared';
-import { UserFavouriteRestaurantsDto } from './UserFavouriteRestaurantsListDto';
+import { ApiController, ApiList, createPaginationLink, Pagination, PaginationQuery, Url } from '../../../shared';
+import { UserFavouriteRestaurantsDto } from './UserFavouriteRestaurantsDto';
 
 @ApiController({
-  path: 'users',
-  name: 'Favourite restaurants',
-  description: 'Operations on favourite restaurants',
+  path: 'users/favourite/restaurants',
+  name: "User's favourite restaurants",
+  description: "Operations on user's favourite restaurants",
 })
 class UserFavouriteRestaurantsController {
   @ApiList({ name: 'restaurants', response: UserFavouriteRestaurantsDto, link: true })
